requirements.txt
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.4
Pillow>=9.5.0
requests>=2.31.0
numpy>=1.24.3
tqdm>=4.65.0
